# agrotech_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AgroTech",
        "version": "1.0.0",
        "title": "AgroTech",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "yogeshwaran@doodlblue.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:2122",
    "basePath": "/api/v1",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/land/getCount": {
            "get": {
                "tags": [
                    "GetCount"
                ],
                "summary": "response",
                "description": "GetCount",
                "operationId": "GetCount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "districtName",
                        "description": "response",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "success": false,
                                "status": 400,
                                "message": "Bad Request"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "example": {
                                "status": 200,
                                "message": "Data Fetched Successfully",
                                "data": {
                                    "totalZoneCount": 2,
                                    "totalVillageCount": 4,
                                    "zoneData": [
                                        {
                                            "zoneName": "Yalamanjili",
                                            "villageCount": 2
                                        },
                                        {
                                            "zoneName": "Yalamanjili1",
                                            "villageCount": 2
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}